+++ src/main/java/com/danigu/blog/ServerApplication.java
@@ -1,6 +1,6 @@
 package com.danigu.blog;
 
-import com.danigu.blog.dic.StationLocator;
+import com.danigu.blog.dic.ServiceLocator;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.servlet.ServletHolder;
@@ -10,7 +10,7 @@ import org.eclipse.jetty.servlet.ServletHolder;
  */
 public class ServerApplication {
     public static void main(String[] args) throws Exception {
-        StationLocator locator = new StationLocator();
+        ServiceLocator locator = new ServiceLocator();
 
         Server server = new Server(8081);
         ServletContextHandler context = new ServletContextHandler(ServletContextHandler.SESSIONS);
@@ -23,7 +23,7 @@ public class ServerApplication {
         server.join();
     }
 
-    public static void configureServlets(StationLocator locator, ServletContextHandler context) {
+    public static void configureServlets(ServiceLocator locator, ServletContextHandler context) {
         BlogServlet blogServlet = new BlogServlet(locator.getPostService(), locator.getCommentService());
         context.addServlet(new ServletHolder(blogServlet), "/post/*");
     }
