+++ src/main/java/com/danigu/blog/post/persistence/PostRepository.java
@@ -1,16 +1,16 @@
 package com.danigu.blog.post.persistence;
 
-import com.danigu.blog.common.persistence.CommonRepository;
+import com.danigu.blog.base.persistence.BaseRepository;
 import com.danigu.blog.post.PostEntity;
 
 import javax.persistence.EntityManagerFactory;
 
 /**
  * DAO for post, it's responsibility to interact with the database.
- * @see CommonRepository
+ * @see BaseRepository
  * @author dani
  */
-public class PostRepository extends CommonRepository<PostEntity> {
+public class PostRepository extends BaseRepository<PostEntity> {
 
     public PostRepository(EntityManagerFactory emf) {
         super(emf, PostEntity.class);
